fix parsing latency and batch size
There was an issue of mismatching scopes when an inner block was being
configured. This was easily resolved by moving a configuration variable
to the class level for the `Routemaster::CLI::Sub::Base` class. Since
the commands will be run individually, class-level configuration changes
wouldn't mean much and they will all be done from scratch for each run.